RPSLS Codebase
Removed unnecessary variables and functions so that only the UI and server/client functionality is recycled

When you connect to a server, you should see that there are 3 "doors". In its current state it's mostly a proof of concept. You click on a door, you get to a new scene which is a new puzzle, complete the puzzle, return to the main menu