Updated "Test Game",
The game plan was originally to have a seperate thread change the scene the players were currently on, so that everyone could be facing off on the same puzzle at the same time. However, I learned that it is currently not feasible to change the scene in a thread that is not the Application thread.

So for now, the paramters of the game are slightly altered. Simply put, players have access to all the puzzles, at any given time. The first player to complete all the puzzles wins the game, at which point the users will be prompted to play again (Play Again feature not yet implemented)