Made different doors instead of one big door, minor changes
Added doors 5 and 6 and all of the helping code to make it work. Moved creating the buttons to leave and submit and creating the textField into its own function to simplify code. Added a try again statement if the input was wrong. Created different doors so that's it's easier to put into the drop down menu. Changed the doors to BorderPanes since all of the doors are no longer in the same arraylist and I don't have to return a VBox with all of the scenes inside of it anymore