Implemented a pop up box in my puzzles
It sort of works, the buttons don't really do what they are supposed to do. If the user presses the pop up button on my puzzles then a dialog box opens in which it prompts the user  asking them if they want to play again or quit. Quit button causes a small core dump, not sure why. Play again button only goes to the scene with the drop down list of buttons, does not disable the door.